// src/api/dto/medical-appointment-view.dto.ts
export interface MedicalAppointmentViewDto {
    id: number;
    reservedId: number;
    reservedPrice: number;
    reservedStatus: 'CONFIRMED' | 'ACTIVE' | 'CANCELED' | 'DELETED';
    doctorRut: string;
    doctorId: number;
    doctorName: string;
    doctorLastName: string;
    schedule: Date;
    officeId: number;
    officeCode: string;
    officeName: string;
    specialityId: number;
    specialityPrice: number;
    specialityCode: string;
    specialityName: string;
    officeFloor: string;
    buildingId: number;
    buildingName: string;
    buildingCode: string;
    centerId: number;
    centerName: string;
    centerCode: string;
    centerAddress: string;
    centerEmail: string;
    centerPhone: string;
    patientId: number;
    patientRut: string;
    patientName: string;
    patientLastName: string;
    patientEmail: string;
}